requirements.txt
python-dotenv
praw
requests
langchain
langchain-groq
langgraph
pandas
tqdm